feat(debug): split tools into per-file structure and add to TOOLS_REQUIRING_NARROWING
- Split monolithic debugTools.ts (1,045 lines) into 9 per-tool files
  under tools/debug/ following the repo's one-file-per-tool convention
- Created shared session-manager.ts for singleton DAP client management
- Added barrel index.ts for clean imports
- Original debugTools.ts now re-exports from debug/ for backward compat
- Added DebugLaunchTool, DebugEvaluateTool, DebugAttachTool to
  TOOLS_REQUIRING_NARROWING for human-in-the-loop security
- Registered DebugAttachTool and DebugSetFunctionBreakpointTool in
  config.ts (were previously missing from tool registration)
