+++ packages/core/src/tools/debug/index.ts
@@ -1,0 +1,22 @@
+/**
+ * @license
+ * Copyright 2026 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Barrel export for all debug tools.
+ *
+ * Each debug tool lives in its own file following the repo's one-file-per-tool
+ * convention. Import from this index to get all tool classes at once.
+ */
+
+export { DebugLaunchTool } from './debug-launch.js';
+export { DebugSetBreakpointTool } from './debug-set-breakpoint.js';
+export { DebugGetStackTraceTool } from './debug-get-stacktrace.js';
+export { DebugGetVariablesTool } from './debug-get-variables.js';
+export { DebugStepTool } from './debug-step.js';
+export { DebugEvaluateTool } from './debug-evaluate.js';
+export { DebugDisconnectTool } from './debug-disconnect.js';
+export { DebugAttachTool } from './debug-attach.js';
+export { DebugSetFunctionBreakpointTool } from './debug-set-function-breakpoint.js';
